// src/game/cities/nigeria-catalogue.generated.ts
export const NIGERIA_CITY_CATALOGUE_ROWS = Object.freeze([
  ["lagos","Lagos","lagos","Lagos State",3.38,6.52,1],
  ["ibadan","Ibadan","oyo","Oyo State",3.95,7.38,1],
  ["abuja","Abuja","fct","Federal Capital Territory",7.49,9.06,1],
  ["port-harcourt","Port Harcourt","rivers","Rivers State",7.03,4.82,1],
  ["abeokuta","Abeokuta","ogun","Ogun State",3.35,7.15,0],
  ["ota","Ota","ogun","Ogun State",3.2,6.68,0],
  ["ijebu-ode","Ijebu-Ode","ogun","Ogun State",3.92,6.82,0],
  ["sagamu","Sagamu","ogun","Ogun State",3.64,6.85,0],
  ["kano","Kano","kano","Kano State",8.52,12,1],
  ["enugu","Enugu","enugu","Enugu",7.5136856,6.4496519,0],
  ["benin-city","Benin City","edo","Edo",5.6221058,6.3330586,1],
  ["calabar","Calabar","cross-river","Cross River",8.3373597,4.9795999,0],
  ["uyo","Uyo","akwa-ibom","Akwa Ibom",7.924791,5.0322447,0],
  ["umuahia","Umuahia","abia","Abia",7.4924165,5.5324032,0],
  ["owerri","Owerri","imo","Imo",7.0341973,5.489736,0],
  ["yenagoa","Yenagoa","bayelsa","Bayelsa",6.2747786,4.9334651,0],
  ["awka","Awka","anambra","Anambra",7.0774372,6.218947,0],
  ["kaduna","Kaduna","kaduna","Kaduna",7.4402777777778,10.523055555556,0],
  ["maiduguri","Maiduguri","borno","Borno",13.15,11.833333333333,0],
  ["gombe","Gombe","gombe","Gombe",11.1667,10.2833,0],
  ["yola","Yola","adamawa","Adamawa",12.483333333333,9.2,0],
  ["ilorin","Ilorin","kwara","Kwara",4.55,8.5,0],
  ["minna","Minna","niger","Niger",6.5569444444444,9.6138888888889,0],
  ["osogbo","Osogbo","osun","Osun",4.5666666666667,7.7666666666667,0],
  ["damaturu","Damaturu","yobe","Yobe",11.966666666667,11.75,0],
  ["jalingo","Jalingo","taraba","Taraba",11.366666666667,8.9,0],
  ["akure","Akure","ondo","Ondo",5.1932647,7.2525595,0],
  ["ado-ekiti","Ado-Ekiti","ekiti","Ekiti",5.22274,7.6232482,0],
  ["asaba","Asaba","delta","Delta",6.7297071,6.1858825,1],
  ["lokoja","Lokoja","kogi","Kogi",6.7441666666667,7.8019444444444,0],
  ["jos","Jos","plateau","Plateau",8.8833333333333,9.9333333333333,0],
  ["lafia","Lafia","nasarawa","Nasarawa",8.5147636,8.4960459,0],
  ["katsina","Katsina","katsina","Katsina",7.6006278,12.9894283,0],
  ["sokoto","Sokoto","sokoto","Sokoto",5.2373194,13.0610588,0],
  ["birnin-kebbi","Birnin Kebbi","kebbi","Kebbi",4.2006114,12.4573462,1],
  ["gusau","Gusau","zamfara","Zamfara",6.6606398,12.1658358,0],
  ["dutse","Dutse","jigawa","Jigawa",9.3390747,11.6909859,0],
  ["bauchi","Bauchi","bauchi","Bauchi",9.8410189,10.3168749,0],
  ["makurdi","Makurdi","benue","Benue",8.5361111111111,7.7305555555556,0],
  ["abakaliki","Abakaliki","ebonyi","Ebonyi",8.1133202,6.3208897,0],
] satisfies readonly CityCatalogueRow[])

export const NIGERIA_CITY_CATALOGUE: readonly CityCatalogueEntry[] = Object.freeze(NIGERIA_CITY_CATALOGUE_ROWS.map(([id, name, stateId, stateName, lon, lat, airport]) => Object.freeze({
  id, name, state: Object.freeze({ id: stateId, name: stateName }), lon, lat, open: true, airport: airport === 1,
})))

export const NIGERIA_CITY_LOADERS: Readonly<Record<string, CityModuleLoader>> = Object.freeze({
  "lagos": async()=>(await import("./lagos/index.ts")).lagosCity,
  "ibadan": async()=>(await import("./ibadan/index.ts")).ibadanCity,
  "abuja": async()=>(await import("./abuja/index.ts")).abujaCity,
  "port-harcourt": async()=>(await import("./port-harcourt/index.ts")).portHarcourtCity,
  "abeokuta": async()=>(await import("./abeokuta/index.ts")).abeokutaCity,
  "ota": async()=>(await import("./ota/index.ts")).otaCity,
  "ijebu-ode": async()=>(await import("./ijebu-ode/index.ts")).ijebuOdeCity,
  "sagamu": async()=>(await import("./sagamu/index.ts")).sagamuCity,
  "kano": async()=>(await import("./kano/index.ts")).kanoCity,
  "enugu": async()=>(await import("./enugu/index.ts")).city,
  "benin-city": async()=>(await import("./benin-city/index.ts")).city,
  "calabar": async()=>(await import("./calabar/index.ts")).city,
  "uyo": async()=>(await import("./uyo/index.ts")).city,
  "umuahia": async()=>(await import("./umuahia/index.ts")).city,
  "owerri": async()=>(await import("./owerri/index.ts")).city,
  "yenagoa": async()=>(await import("./yenagoa/index.ts")).city,
  "awka": async()=>(await import("./awka/index.ts")).city,
  "kaduna": async()=>(await import("./kaduna/index.ts")).city,
  "maiduguri": async()=>(await import("./maiduguri/index.ts")).city,
  "gombe": async()=>(await import("./gombe/index.ts")).city,
  "yola": async()=>(await import("./yola/index.ts")).city,
  "ilorin": async()=>(await import("./ilorin/index.ts")).city,
  "minna": async()=>(await import("./minna/index.ts")).city,
  "osogbo": async()=>(await import("./osogbo/index.ts")).city,
  "damaturu": async()=>(await import("./damaturu/index.ts")).city,
  "jalingo": async()=>(await import("./jalingo/index.ts")).city,
  "akure": async()=>(await import("./akure/index.ts")).city,
  "ado-ekiti": async()=>(await import("./ado-ekiti/index.ts")).city,
  "asaba": async()=>(await import("./asaba/index.ts")).city,
  "lokoja": async()=>(await import("./lokoja/index.ts")).city,
  "jos": async()=>(await import("./jos/index.ts")).city,
  "lafia": async()=>(await import("./lafia/index.ts")).city,
  "katsina": async()=>(await import("./katsina/index.ts")).city,
  "sokoto": async()=>(await import("./sokoto/index.ts")).city,
  "birnin-kebbi": async()=>(await import("./birnin-kebbi/index.ts")).city,
  "gusau": async()=>(await import("./gusau/index.ts")).city,
  "dutse": async()=>(await import("./dutse/index.ts")).city,
  "bauchi": async()=>(await import("./bauchi/index.ts")).city,
  "makurdi": async()=>(await import("./makurdi/index.ts")).city,
  "abakaliki": async()=>(await import("./abakaliki/index.ts")).city,
})